Indent input mode image and timer text properly in portrait mode.
Around the grid player view a margin is reserved to display the swipe border. The width of this border is dependent on the grid size. A customized relative layout is added which is able to increase specified margins to the width of the swipe border. As a result the controls in this relative layout can be properly aligned with the visible grid.