Prevent null pointer exceptions when destroying fragments.
On resuming the app sometimes a null pointer exception occurred. This was especially the case when other apps were opened and used without closing the MathDoku app. In case the app was destroyed by Android in order to reuse the memory this could lead to a null pointer exception in the onDestroy calls. This behavior can be faked by enabling Setting -> Developer options -> Don't keep activities.