Pause and resume when day dreaming (API17+) starts and stop.
Android 17+ (Jelly Bean MR1) introduced the day dream mode. In case the MathDoku setting "keep display on" is disabled and Android's day dreaming is turned on, the timer was not stopped as soon as day dreaming started. Now MathDoku responds correctly on the intent broadcasts for start and stopping day dreaming.