Restore the 3x3 keypad for the digit buttons.
Dependent on the settings the 3x3 key will be entire displayed in both
portrait as landscape mode. If needed the grid will be resized (e.g. made
smaller so that the full grid can be displayed).
Also the layout of the buttons is changed to a dark background. Digits on
the buttons are painted in the input mode color (unless the corresponding
setting has been disabled).