Move all basic (non swipe related) to new GridBasePlayerView class.
The class GridPlayerView has been splitted into a new class
GridBasePlayerView which contain all basic operations to play a puzzle
without swiping functionality. Technically this class suffies to be played
with digit buttons only.

All swipe related stuff is kept into the GridPlayerView class.

Method markInvalidChoices() is moved to the Grid class as it was mainly
not UI related.

Finally a bug was solved when check progress was used as each invalid
value was counted twice.
related