Display ticker tape and timer in same area.
The timer is now displayed centered above the grid. In case the ticker
tape has to be shown it will temporarily cover the timer. This increases
the total height available for displaying the grid.

Horizontal aligning the timer text was needed to work around a layout
problem with the grid. In the old situation the grid was not properly
aligned horizontally as a result of aligning the timer text with the
visible grid on the right side.

Note: some work also needs to be done on the input mode button which now
partially covers the grid.