Google play services resource id xml now depents on build command.
Dependent on whether "ant debug" or "ant release" is used to build the
app, the correct files with resource id for the google play services is
copied to the resources folder.

The config folder has been restructured to offer more generic support for
templates. Folder structure of this folder has to correspond with the
folder structure of the target as files will be copied to the same folder
as is used in the ANT-templates folder.

Files in the source folders (src and res) which will be replaced by the
ANT-build scripts have been put in the gitignore.