Reset motion in base player view on load of new grid.
When starting new games of different sizes it is necessary to reset the
motion so the changed grid view properties are synched with the motion
class. This resolves a bug in which cells could not properly be selected
after starting a new game of another (higher) grid size while the current
puzzle was not yet finished.