Replace "Play again" with "Play a similar game" after finishing a puzzle.
The archive fragment now optionally displays a big action button in the
content area.

In case the fragment is displayed in the puzzle activity when a puzzle is
just completed, the button "Play a similar game" will be shown. This will
create a new game with exact same puzzle parameter settings as were used
for the previous game. The replay function is still available via the
menu.

In case the archive fragment is displayed by the archive activity the
action button "Continue this puzzle" will be displayed for an unfinished
puzzle only. For a finished puzzle, no action button is shown. The replay
function is available via the menu.

When choosing to replay a game, a confirmation dialog is first shown.