Refactor submitting scores.
As the number of leaderboards was increased, the overhead of retrieving
all leaderboards after signin to Google (which happened on each resume of
the puzzle fragment activity) was getting too high. The refactor has
reduced this to a minimum. For this the new table leaderboard_rank is
added to the database. This table holds per leaderboard the best score as
registered on Google Play Services. In case a new best score is achieved
the leaderboard rank table is being updated and the Google Play Service
Rank is deleted. After that a process is started to update the ranking
information for the leaderboard.

On each signin to Google+ it now only has to checked for which
leaderboards the ranking information is missing. For those leaderboards
the score is submitted again.

For testing purposes, a development helper is added to manually submit a
score to the Development Leaderboards. To achieve this, the method
unrevealSolution was added to class grid and the method getGrid was added
to the ArchiveFragment.