Initialize all (new) leaderboards after sign in to Google+.
A new preference is used to indicate whether new leaderboards have been
added in an app-update. If the preference indicates that new leaderboards
have been added than those leaderboards are added to the leaderboard rank
table in the database upon sign in to Google+.

The leaderboard rank table now holds information which can be used to
detect the leaderboards which needs to be updated (e.g. new leaderboards
and or leaderboards for which a new local top score has been achieved).