Prevent multiple fragment being initialized at same time.
On a Sony Xperia E it happened that the replay and share action buttons
were shown on an active puzzle in case a new puzzle was generated while
the ArchiveFragment was displayed. This error was caused because the
new fragment was created before the reference to the old fragment was
cleared. As a result both fragment references were not null while
evaluating which action button have to be displayed. Probably this
happened on other slower (single core) devices as well.