Display the leaderboard overview after manual sign in to Google+.
In case the leaderboards exists which need to be updated when manual sign
to Google+ is completed, a progress dialog is shown. Upon completion of
updating the leaderboards, the leaderboard overview is displayed.

When navigating to the leaderboard overview the update process of the
outdated leaderboards is continued if not yet completed. The tab which is
displayed will however not be updated until the fragment is created again.