Add the first (failing) unit test to the project.
The build files have been restructured in order to be able to add unit
tests to the project. Both building in Android Studio and from the command
line is possible.
Currently one failing unit test has been added as a proof that builds do
not succeed in case of a failing unit test.