Display border after creating a new grid.
After creating a new Grid the cage borders were not displayed until a cell
in the cage was selected, or the entire grid was loaded again. This error
was caused when refactoring GridCage.checkCageMathCorrect. The borders are
now redrawn when calling the method on a newly created cage by defaulting
the mUserMathCorrect to false. As for a new cage the cage math is always
true, the borders will be redrawn when first calling this method.