Create new grid in setup.
For most test the grid created in method setup can be used. The
ObjectsCreator for this grid creates mocks for all objects and empty
ArrayLists for the cells, cages and cell changes.

For some tests, this default grid does not suffice for testing. In such
case the ObjectCreator of the Grid itself is recreated.