Unit test no longer fails on "gradle build".
The unit tests succeeded when running "gradle check" as the unit tests were
only ran with the debug sources. When running "gradle build" the unit tests
are executed with both debug as wel as with the release sources. The tests
with the release sources failed although the same tests with the debug
sources succeeded due to different results of method Grid.unrevealSolution.