Move fromStorageString and toStorageString from GridCell to new class
The new class GridCellStorage is responsible for converting a GridCell to
a string which can be persisted. Also this class takes care of parsing
the storage string. The GridCellStorage object can be passed to a special
GridCell constructor to instantiate a new grid cell with the values of
the GridCellStorage object.