Refactor Grid.Replay to Grid.createNewGridForReplay.
Instead of clearing the existing grid, an entire new grid is instantiated
using the Grid Builder.

Also refactored class TestData to TestGrid and subclasses. The test classes
are use when it is not possible to use the mocked versions. It should be
considered to use these test classes more often instead of the current
GridBuilderStub class in the GridTest as they are easier to understand
and maintain.