Add unit tests for class GridGenerator.
To unit test this class, a few puzzles have been generated with the app.
Those puzzles have been hard coded in the GridCreator classes. All grid
generating parameters which were used are also stored in these classes.
Finally the grid definition for this puzzle was deferred and stored as
hard coded string in the classes.

To test the grid generation, the grids are generated with the stored grid
generating parameters. Next the grid definition of the generated grid is
compared with the hard coded grid definition in the grid creator classes.

Following grid creator classes can be used to test the grid generator:
* Grid2x2
* Grid4x4HiddenOperators
* Grid9x9

The other grid creators can not be used to test the grid generators as
these have not been generated entirely with the app.

Also included in this commit:
* The grid definition now renumber the cage id's so all cage ids are
  consecutive, starting from 0. Also the top left cells of all cages are
  now sorted in order of ascending cell numbers. In this way, grids having
  the exact same cage pattern and solution but just different cage id's
  will now be recognized as identical.
* Unit tests which used the old TestGrids are now refactored to use the
  new grid creators.