Add a development option to generate the source code for the GridCreators.
Each time the grid generating algorithm is changed, the test helper
classes which create the test grids (GridCreators) get invalid. As it
very error sensitive to create the data for the classes manually, the code
for these classes can now be generated manually when running in
development mode. The generated code is written to the debug log.