The first (bigger) cage is placed at the unshifted position.
If applicable the cage matrix starts with a bigger cage at a random
position in the grid. As the cage type is placed at a random position,
there is no need to shift the cage type in case the top left cell of the
cage type does not start at the top left cell of the retangle which
encloses the cage type.