Update elapsed time in grid only when timer was start for this grid.
This is a bug in the current production version (revision 594, version
2.1.0) of the app.

Previously, the timer was not resetted when a new puzzle is generated
while the current grid was not yet finished. In the new situation the id
of the solving attempt for which the timer is created is stored within
the timer. When stopping the timer, this id is used to check whether the
elapsed time should be updated in the solving attempt.