Refactor test.
These tests seemed to regularly to break when just another test class was
changed. After a Gradle clean and full rebuild the error disappeared.
Reason for this could be the improrer handling of the databaseHelper
variable. Now the existing databaseHelper is closed before it is replace
with a stubbed version of the DatabaseHelper.